pyproject.toml
[project]
name = "p2-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic[bedrock]>=0.51.0",
    "faker>=37.3.0",
    "fastapi>=0.115.12",
    "fitz>=0.0.1.dev2",
    "ipykernel>=6.29.5",
    "numpy>=2.2.5",
    "ollama>=0.4.8",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.6",
    "pymupdf>=1.25.5",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "reportlab>=4.4.0",
    "spacy>=3.8.5",
    "uvicorn>=0.34.2",
]
